+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this
 file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- Changed assetsForAddress() to assetsForAddressWithLimit() to enforce asset limit in actions package ([#156](https://github.com/stellar/stellar-horizon/pull/156)).
+
 ## 25.0.1
 
 ### Fixed